+++ chat/src/main/java/chat/application/ChatRoomSaveService.java
@@ -2,13 +2,11 @@ package chat.application;
 
 import chat.adapter.output.client.ArticleClient;
 import chat.adapter.output.client.UserClient;
-import chat.adapter.output.client.response.ArticleFeignResponse;
-import chat.application.port.input.ChatRoomCheckUseCase;
+import chat.adapter.output.client.dto.ArticleFeignInfo;
+import chat.application.port.input.UserChatSaveUseCase;
 import chat.application.port.output.ChatRoomPersistencePort;
-import chat.application.port.output.UserChatPersistencePort;
 import chat.domain.command.ChatRoomSaveCommand;
 import chat.domain.dto.ChatRoomSaveForm;
-import chat.domain.dto.UserChatSaveForm;
 import java.util.List;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
@@ -19,11 +17,9 @@ import org.springframework.stereotype.Service;
 @RequiredArgsConstructor
 public class ChatRoomSaveService {
 
-    private final ChatRoomCheckUseCase chatRoomCheckUseCase;
-
-    private final UserChatPersistencePort userChatPersistencePort;
-
     private final ChatRoomPersistencePort chatRoomPersistencePort;
+
+    private final UserChatSaveUseCase userChatSaveUseCase;
 
     private final ChatRoomGenerator chatRoomGenerator;
 
@@ -33,7 +29,7 @@ public class ChatRoomSaveService {
 
     public String saveChatRoom(ChatRoomSaveCommand chatRoomSaveCommand) {
         String buyerId = chatRoomSaveCommand.getBuyerId();
-        ArticleFeignResponse articleInfo = getArticleBy(chatRoomSaveCommand.getArticleId());
+        ArticleFeignInfo articleInfo = getArticleBy(chatRoomSaveCommand.getArticleId());
 
         // articleId로 ChatRoom 존재 여부 확인, 없으면 생성
         String chatRoomId = chatRoomPersistencePort.getChatRoomBy(chatRoomSaveCommand.getArticleId())
@@ -42,17 +38,17 @@ public class ChatRoomSaveService {
             );
 
         // UserChat 존재 유뮤 확인, 없으면 생성
-        saveUserChatIfNotExists(chatRoomId, buyerId);
-        saveUserChatIfNotExists(chatRoomId, articleInfo.getUserId());
+        userChatSaveUseCase.saveUserChatIfNotExists(chatRoomId, buyerId);
+        userChatSaveUseCase.saveUserChatIfNotExists(chatRoomId, articleInfo.getUserId());
 
         return chatRoomId;
     }
 
-    private ArticleFeignResponse getArticleBy(String articleId) {
+    private ArticleFeignInfo getArticleBy(String articleId) {
         return articleClient.getArticleBy(articleId);
     }
 
-    private String createChatRoom(String buyerId, String articleId, ArticleFeignResponse articleInfo) {
+    private String createChatRoom(String buyerId, String articleId, ArticleFeignInfo articleInfo) {
         String buyerNickName = userClient.getNickname(buyerId);
         String sellerNickName = userClient.getNickname(articleInfo.getUserId());
 
@@ -61,13 +57,6 @@ public class ChatRoomSaveService {
 
         return chatRoomPersistencePort.saveChatRoom(
             ChatRoomSaveForm.of(title, articleId, articleInfo.getArticleImage()));
-    }
-
-    private void saveUserChatIfNotExists(String chatRoomId, String userId) {
-        Boolean existsUserChat = userChatPersistencePort.existsBy(chatRoomId, userId);
-        if (!existsUserChat) {
-            userChatPersistencePort.saveUserChat(UserChatSaveForm.of(chatRoomId, userId));
-        }
     }
 
 }
