+++ chat/src/main/java/chat/application/port/input/UserChatSaveUseCase.java
@@ -1,0 +1,7 @@
+package chat.application.port.input;
+
+public interface UserChatSaveUseCase {
+
+    void saveUserChatIfNotExists(String chatRoomId, String userId);
+
+}
